flask-api/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Serialization and validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Kafka
kafka-python==2.0.2

# Monitoring
prometheus-client==0.19.0

# HTTP server
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0

# Testing (development)
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
faker==20.1.0
